refactor: revert loader store
Revert loader store
