--- src/exporter/models/loaderStore.ts
@@ -1,146 +1,0 @@
-/* eslint-disable camelcase */
-import { types, Instance, flow, getParent } from 'mobx-state-tree';
-import { AxiosRequestConfig, AxiosError } from 'axios';
-import { ApiHttpResponse } from '../../common/models/api-response';
-import { ResponseState } from '../../common/models/ResponseState';
-import { LoaderStoreError } from '../../common/models/loaderStoreError';
-import { searchParams } from './search-params';
-import { IRootLoaderStore } from './rootLoaderStore';
-import { ILoadModelStatus } from './loadModelStatus';
-
-export type LoadModelStatusResponse = ILoadModelStatus[];
-export interface LoadResult {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: any;
-}
-
-export interface ModelInfo {
-  modelPath: string;
-  tilesetFilename: string;
-  identifier: string;
-}
-
-export type LoaderResponse = ApiHttpResponse<LoadResult>;
-
-const internalError = types.model({
-  request: types.maybe(types.frozen<AxiosRequestConfig>()),
-  key: types.frozen<LoaderStoreError>(),
-});
-export interface IInternalError extends Instance<typeof internalError> {}
-
-export const loaderStore = types
-  .model({
-    state: types.enumeration<ResponseState>(
-      'State',
-      Object.values(ResponseState)
-    ),
-    searchParams: types.optional(searchParams, {}),
-    models: types.maybe(types.frozen<any>([])),
-    errors: types.frozen<IInternalError[]>([]),
-  })
-  .views((self) => ({
-    get root(): IRootLoaderStore {
-      return getParent(self);
-    },
-  }))
-  .actions((self) => {
-    const startLoadModel: (
-      modelInfo: ModelInfo
-    ) => Promise<void> = flow(function* startLoadModel(
-      modelInfo: ModelInfo
-    ): Generator<Promise<LoaderResponse>, void, LoaderResponse> {
-      self.state = ResponseState.PENDING;
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const params: Record<string, unknown> = {};
-      // Prepare body data for request
-      params.modelPath = modelInfo.modelPath;
-      params.tilesetFilename = modelInfo.tilesetFilename;
-      params.identifier = modelInfo.identifier;
-
-      try {
-        const result = yield self.root.fetch(
-          '/models',
-          'POST',
-          params
-        );
-        // const responseBody = result.data.data;
-        self.state = ResponseState.DONE;
-      } catch (e) {
-        const error = e as AxiosError;
-        // eslint-disable-next-line
-        if (error) {
-          // eslint-disable-next-line
-          if (error.response && error.response.data && error.response.data.name) {
-            addError({
-              request: error.config,
-              // eslint-disable-next-line
-              key: error.response.data.name as LoaderStoreError,
-            });
-          } else {
-            addError({
-              request: error.config,
-              key: LoaderStoreError.GENERAL_ERROR,
-            });
-          }
-        }
-        self.state = ResponseState.ERROR;
-      }
-    });
-    const getModels: () => Promise<void> = flow(
-      function* getModels(): Generator<
-        Promise<LoaderResponse>,
-        void,
-        LoadModelStatusResponse
-      > {
-        try {
-          self.state = ResponseState.IDLE;
-          const result = yield self.root.fetch('/jobs', 'GET', {});
-          self.models = result;
-        } catch (e) {
-          const error = e as AxiosError;
-          self.state = ResponseState.ERROR;
-          addError({
-            request: error.config,
-            key: LoaderStoreError.GENERAL_ERROR,
-          });
-        }
-      }
-    );
-
-    const addError = (error: IInternalError): void => {
-      self.errors = [...self.errors, error];
-    };
-
-    const hasError = (key: LoaderStoreError): boolean => {
-      return Boolean(self.errors.find((err) => err.key === key));
-    };
-
-    const hasErrors = (): boolean => {
-      const minimalLength = 0;
-      return self.errors.length > minimalLength;
-    };
-
-    const cleanError = (key: LoaderStoreError): boolean => {
-      if (hasError(key)) {
-        self.errors = self.errors.filter((err) => err.key !== key);
-        return true;
-      }
-      return false;
-    };
-
-    const cleanErrors = (): void => {
-      self.errors = [];
-    };
-
-    return {
-      startLoadModel,
-      getModels,
-      addError,
-      hasError,
-      hasErrors,
-      cleanError,
-      cleanErrors,
-    };
-  });
-
-export interface IConflictsLoaderStore extends Instance<typeof loaderStore> {}
